+++ src/modules/Elsa.Caching.Distributed.MassTransit/Services/MassTransitChangeTokenSignalPublisher.cs
@@ -1,4 +1,5 @@
 ﻿using Elsa.Caching.Contracts;
+using Elsa.Caching.Distributed.Contracts;
 using Elsa.Caching.Distributed.MassTransit.Messages;
 using MassTransit;
