+++ src/modules/Elsa.Caching.Distributed/Services/DistributedChangeTokenSignaler.cs
@@ -1,14 +1,15 @@
 using Elsa.Caching.Contracts;
+using Elsa.Caching.Distributed.Contracts;
 using JetBrains.Annotations;
 using Microsoft.Extensions.Primitives;
 
-namespace Elsa.Caching.Services;
+namespace Elsa.Caching.Distributed.Services;
 
 /// <summary>
 /// Decorates an <see cref="IChangeTokenSignaler"/> and publishes a signal after the signal has been triggered.
 /// </summary>
 [UsedImplicitly]
-public class DistributedChangeTokenSignaler(IChangeTokenSignaler decoratedSignaler, IChangeTokenSignalPublisher signalPublisher) : IChangeTokenSignaler, IDistributedChangeTokenSignaler
+public class DistributedChangeTokenSignaler(IChangeTokenSignaler decoratedSignaler, IChangeTokenSignalPublisher signalPublisher) : IChangeTokenSignaler, IChangeTokenSignalInvoker
 {
     /// <inheritdoc />
     public IChangeToken GetToken(string key)
@@ -19,12 +20,12 @@ public class DistributedChangeTokenSignaler(IChangeTokenSignaler decoratedSignal
     /// <inheritdoc />
     public async ValueTask TriggerTokenAsync(string key, CancellationToken cancellationToken = default)
     {
-        await TriggerTokenLocalAsync(key, cancellationToken);
+        await decoratedSignaler.TriggerTokenAsync(key, cancellationToken);
         await signalPublisher.PublishAsync(key, cancellationToken);
     }
 
     /// <inheritdoc />
-    public async ValueTask TriggerTokenLocalAsync(string key, CancellationToken cancellationToken = default)
+    async ValueTask IChangeTokenSignalInvoker.InvokeAsync(string key, CancellationToken cancellationToken)
     {
         await decoratedSignaler.TriggerTokenAsync(key, cancellationToken);
     }
