+++ src/modules/Elsa.Caching/Extensions/ModuleExtensions.cs
@@ -10,27 +10,10 @@ namespace Elsa.Extensions;
 public static class ModuleExtensions
 {
     /// <summary>
-    /// Adds the <see cref="DistributedCacheFeature"/> feature to the system.
+    /// Adds the <see cref="MemoryCacheFeature"/> feature to the system.
     /// </summary>
     public static MemoryCacheFeature UseMemoryCache(this IModule module, Action<MemoryCacheFeature>? configure = default)
     {
         return module.Configure(configure);
-    }
-    
-    /// <summary>
-    /// Adds the <see cref="DistributedCacheFeature"/> feature to the system.
-    /// </summary>
-    public static DistributedCacheFeature UseDistributedCache(this MemoryCacheFeature memoryCacheFeature, Action<DistributedCacheFeature>? configure = default)
-    {
-        return memoryCacheFeature.Module.Configure(configure);
-    }
-    
-    /// <summary>
-    /// Adds the <see cref="DistributedCacheFeature"/> feature to the system.
-    /// </summary>
-    public static IModule UseDistributedCache(this IModule module, Action<DistributedCacheFeature>? configure = default)
-    {
-        module.Configure(configure);
-        return module;
     }
 }
